+++ packages/form/src/demos/layout-change.tsx
@@ -57,7 +57,7 @@ export default () => {
           onFinish={async (values: any) => {
             await waitTime(2000);
             console.log(values);
-            message.success('提交成功！');
+            message.success('提交成功');
           }}
           initialValues={{
             name: '蚂蚁设计有限公司',