+++ packages/table/src/utils.tsx
@@ -6,14 +6,12 @@ import {
   isNil,
   LabelIconTip,
   omitUndefinedAndEmptyArr,
+  ProFieldValueType,
   ProSchemaComponentTypes,
   ProTableEditableFnType,
+  UseEditableUtilType,
 } from '@ant-design/pro-utils';
-import {
-  ProFieldEmptyText,
-  proFieldParsingValueEnumToArray,
-  ProFieldValueType,
-} from '@ant-design/pro-field';
+import { ProFieldEmptyText, proFieldParsingValueEnumToArray } from '@ant-design/pro-field';
 import get from 'rc-util/lib/utils/get';
 import { IntlType } from '@ant-design/pro-provider';
 
@@ -25,9 +23,7 @@ import {
   UseFetchDataAction,
 } from './typing';
 import { ColumnsState, useCounter } from './container';
-import defaultRenderText, { spellNamePath } from './defaultRender';
-import { UseEditableUtilType } from './component/useEditable';
-import InlineErrorFormItem from './component/InlineErrorFormItem';
+import defaultRenderText from './defaultRender';
 
 /**
  * 检查值是否存在
@@ -288,7 +284,8 @@ export function columnRender<T>({
   const { renderText = (val: any) => val } = columnProps;
 
   const renderTextStr = renderText(text, rowData, index, action.current as ActionType);
-
+  const mode =
+    isEditable && !isEditableCell(text, rowData, index, columnProps?.editable) ? 'edit' : 'read';
   const textDom = defaultRenderText<T>({
     text: renderTextStr,
     valueType: (columnProps.valueType as ProFieldValueType) || 'text',
@@ -298,17 +295,18 @@ export function columnRender<T>({
     columnEmptyText,
     type,
     recordKey,
-    mode: isEditable ? 'edit' : 'read',
+    mode,
   });
 
-  const dom: React.ReactNode = isEditable
-    ? textDom
-    : genEllipsis(genCopyable(textDom, columnProps, renderTextStr), columnProps, renderTextStr);
+  const dom: React.ReactNode =
+    mode === 'edit'
+      ? textDom
+      : genEllipsis(genCopyable(textDom, columnProps, renderTextStr), columnProps, renderTextStr);
 
   /**
    * 如果是编辑模式，并且 renderFormItem 存在直接走 renderFormItem
    */
-  if (isEditable) {
+  if (mode === 'edit') {
     if (columnProps.valueType === 'option') {
       return (
         <Form.Item shouldUpdate noStyle>
@@ -326,44 +324,7 @@ export function columnRender<T>({
         </Form.Item>
       );
     }
-    if (isEditableCell(text, rowData, index, columnProps?.editable)) {
-      return (
-        <Form.Item shouldUpdate noStyle>
-          {text}
-        </Form.Item>
-      );
-    }
-    if (columnProps.renderFormItem) {
-      return (
-        <Form.Item shouldUpdate noStyle>
-          {(form: any) => {
-            const inputDom = columnProps.renderFormItem?.(
-              {
-                ...columnProps,
-                isEditable: true,
-              },
-              {
-                defaultRender: () => <>{dom}</>,
-                type,
-              },
-              form,
-            );
-            return (
-              <InlineErrorFormItem
-                initialValue={text}
-                name={spellNamePath(
-                  recordKey || index,
-                  columnProps?.key || columnProps?.dataIndex || index,
-                )}
-                {...columnProps.formItemProps}
-              >
-                {inputDom || dom}
-              </InlineErrorFormItem>
-            );
-          }}
-        </Form.Item>
-      );
-    }
+    return dom;
   }
 
   if (columnProps.render) {
@@ -378,6 +339,7 @@ export function columnRender<T>({
       {
         ...columnProps,
         isEditable,
+        type: 'table',
       },
     );